+++ .vscode/settings.json
@@ -51,5 +51,5 @@
         "temp/": true,
         "Temp/": true
     },
-    "dotnet.defaultSolution": "Pirate Game - Comp6100.sln"
+    "dotnet.defaultSolution": "Mobile-Pirate-Game.sln"
 }
