+++ Assets/GameAssets/Prefab/CannonLevel/6CBStack.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 71ba5a1cfad83134380ee731c24ed1d1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
