+++ Assets/GameAssets/Sprites/CannonLevel/seaugull flying.png.meta
